requirements.txt
# Core dependencies
asyncio-mqtt==0.16.1
pydantic==2.5.0
pyyaml==6.0.1
httpx==0.25.0
aiohttp==3.9.0

# Database
asyncpg==0.29.0
redis==5.0.1
sqlalchemy[asyncio]==2.0.23

# Metrics & Monitoring  
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
mypy==1.7.0
ruff==0.1.6
